refactor(cockpit-render): move json-highlight tokenizer to render/shared for reuse
