+++ cockpit/render/shared/json-highlight.spec.ts
