+++ cockpit/render/shared/json-highlight.ts
